Fixes #271
