+++ mesa/agent.py
@@ -1,0 +1,19 @@
+# -*- coding: utf-8 -*-
+"""
+The agent class for Mesa framework.
+
+Core Objects: Agent
+
+"""
+
+
+class Agent:
+    """ Base class for a model agent. """
+    def __init__(self, unique_id, model):
+        """ Create a new agent. """
+        self.unique_id = unique_id
+        self.model = model
+
+    def step(self, model):
+        """ A single step of the agent. """
+        pass
